The Principle of Last Privilege : Update StoreTicketRequest, UpdateTicketRequest, TicketController, and ApiController;

+++ app/Http/Controllers/Api/V1/TicketController.php
@@ -30,15 +30,13 @@ class TicketController extends ApiController
      */
     public function store(StoreTicketRequest $request)
     {
-       try {
-            // Policy
-            $this->isAble('store', Ticket::class);
-
+        // Policy
+        if ($this->isAble('store', Ticket::class)) {
             // Create ticket
             return new TicketResource(Ticket::create($request->mappedAttributes()));
-        } catch (AuthorizationException $ex) {
-            return $this->error('You are not authorized to create that resource.', 401);
         }
+
+        return $this->error('You are not authorized to create that resource.', 401);
     }
 
     /**
@@ -69,15 +67,16 @@ class TicketController extends ApiController
             $ticket = Ticket::findOrFail($ticket_id);
 
             // Policy
-            $this->isAble('update', $ticket);
+            if ($this->isAble('update', $ticket)) {
+                $ticket->update($request->mappedAttributes());
 
-            $ticket->update($request->mappedAttributes());
+                return new TicketResource($ticket);
+            }
 
-            return new TicketResource($ticket);
+            return $this->error('You are not authorized to update that resource.', 401);
+
         } catch (ModelNotFoundException $exception) {
             return $this->error('Ticket cannot be found.', 404);
-        } catch (AuthorizationException $ex) {
-            return $this->error('You are not authorized to update that resource.', 401);
         }
     }
 
@@ -88,7 +87,11 @@ class TicketController extends ApiController
             $ticket = Ticket::findOrFail($ticket_id);
 
             // Policy
-            $this->isAble('replace', $ticket);
+            if ($this->isAble('replace', $ticket)) {
+                $ticket->update($request->mappedAttributes());
+
+                return new TicketResource($ticket);
+            }
 
             // $model = [
             //     'title' => $request->input('data.attributes.title'),
@@ -98,9 +101,8 @@ class TicketController extends ApiController
             // ];
             // $ticket->update($model);
 
-            $ticket->update($request->mappedAttributes());
+            return $this->error('You are not authorized to replace that resource.', 401);
 
-            return new TicketResource($ticket);
         } catch (ModelNotFoundException $exception) {
             return $this->error('Ticket cannot be found.', 404);
         }
@@ -115,11 +117,14 @@ class TicketController extends ApiController
             $ticket = Ticket::findOrFail($ticket_id);
 
             // Policy
-            $this->isAble('delete', $ticket);
+            if ($this->isAble('delete', $ticket)) {
+                $ticket->delete();
 
-            $ticket->delete();
+                return $this->ok('Ticket deleted successfully');
+            }
 
-            return $this->ok('Ticket deleted successfully');
+            return $this->error('You are not authorized to delete that resource.', 401);
+
         } catch (ModelNotFoundException $exception) {
             return $this->error('Ticket not found.', 404);
         }
